fix: prevent integer overflow in auxiliary loss normalization calculations
GraphConvolutionalLayer.cs:
- Cast batchSize and _graphEdges.Count to long before multiplication
- Prevents overflow when computing totalEdges for large graphs

SqueezeAndExcitationLayer.cs:
- Cast Rows and Columns to long before multiplication
- Prevents overflow when computing totalElements for large attention matrices

Integer overflow can occur when multiplying two large int values before converting to double, resulting in incorrect loss normalization.

Addresses PR #474 comments on potential overflow in auxiliary loss computations.

+++ src/NeuralNetworks/Layers/GraphConvolutionalLayer.cs
@@ -726,7 +726,7 @@ public class GraphConvolutionalLayer<T> : LayerBase<T>
         }
 
         // Normalize by batch size and number of edges for scale-invariance
-        T totalEdges = NumOps.FromDouble(batchSize * _graphEdges.Count);
+        T totalEdges = NumOps.FromDouble((long)batchSize * (long)_graphEdges.Count);
         smoothnessLoss = NumOps.Divide(smoothnessLoss, totalEdges);
 
         // Apply smoothness weight to allow tuning of this auxiliary loss
